requirements.txt
scapy
requests
colorama
pyyaml